hantek-dso: Capture ratio is a uint64_t.
From src/hwdriver.c:

  {SR_CONF_CAPTURE_RATIO, SR_T_UINT64, "captureratio",
          "Pre-trigger capture ratio", NULL},

+++ src/hardware/hantek-dso/protocol.h
@@ -204,7 +204,7 @@ struct dev_context {
 	gboolean filter[2];
 	int triggerslope;
 	char *triggersource;
-	int capture_ratio;
+	uint64_t capture_ratio;
 	int triggermode;
 
 	/* Frame transfer */
